SOV-3767: Portfolio refactoring
