+++ apps/frontend/src/app/5_pages/PortfolioPage/PortfolioPage.tsx
@@ -14,7 +14,7 @@ const PortfolioPage: FC = () => (
     </Helmet>
 
     <div className="px-0 container md:mx-9 mx-0 md:mb-2 mt-4 md:mt-12 mb-7">
-      <div className="grid grid-cols-9 gap-10 md:gap-12 lg:gap-20 w-full mb-6">
+      <div className="grid grid-cols-9 items-start gap-10 md:gap-12 lg:gap-20 w-full mb-6">
         <div className="col-span-9 md:col-span-4 lg:col-span-3 md:bg-gray-90 rounded md:py-7 md:px-6">
           <ProtocolSection />
         </div>
