+++ apps/frontend/src/app/5_pages/PortfolioPage/components/AssetSection/components/AssetSectionActions/AssetSectionActions.tsx
@@ -9,12 +9,14 @@ import { Button, ButtonStyle, Paragraph } from '@sovryn/ui';
 import { ReactComponent as ConvertIcon } from '../../../../../../../assets/images/convert.svg';
 import { ReactComponent as DepositIcon } from '../../../../../../../assets/images/deposit.svg';
 import { ReactComponent as WithdrawIcon } from '../../../../../../../assets/images/withdraw.svg';
+import { useAccount } from '../../../../../../../hooks/useAccount';
 import { useAssetBalance } from '../../../../../../../hooks/useAssetBalance';
 import { translations } from '../../../../../../../locales/i18n';
 import { sharedState } from '../../../../../../../store/rxjs/shared-state';
 
 export const AssetSectionActions: FC = () => {
   const navigate = useNavigate();
+  const { account } = useAccount();
 
   const { balance } = useAssetBalance(SupportedTokens.rbtc);
   const hasRbtcBalance = useMemo(() => Number(balance) !== 0, [balance]);
@@ -31,7 +33,7 @@ export const AssetSectionActions: FC = () => {
 
   return (
     <>
-      <div className="flex md:hidden w-full justify-center items-center gap-8 my-4">
+      <div className="flex md:hidden w-full justify-center items-center gap-8">
         <div className="flex flex-col gap-3 items-center">
           <button
             onClick={handleFundWallet}
@@ -54,7 +56,7 @@ export const AssetSectionActions: FC = () => {
             {t(translations.portfolioPage.assetSection.convert)}
           </Paragraph>
         </div>
-        {hasRbtcBalance && (
+        {hasRbtcBalance && account && (
           <div className="flex flex-col gap-3 items-center">
             <button
               onClick={handleWithdraw}
@@ -68,7 +70,7 @@ export const AssetSectionActions: FC = () => {
           </div>
         )}
       </div>
-      <div className="hidden md:flex items-center gap-4 my-6">
+      <div className="hidden md:flex items-center gap-4 mt-2.5 mb-6">
         <Button
           className="w-[7.75rem]"
           style={ButtonStyle.primary}
@@ -81,7 +83,7 @@ export const AssetSectionActions: FC = () => {
           onClick={() => navigate('/convert?&to=sov')}
           text={t(translations.portfolioPage.assetSection.convert)}
         />
-        {hasRbtcBalance && (
+        {hasRbtcBalance && account && (
           <Button
             className="w-[7.75rem]"
             style={ButtonStyle.secondary}
