+++ apps/frontend/src/app/5_pages/PortfolioPage/components/ProtocolSection/components/ProtocolTotalSection/ProtocolTotalSection.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useMemo, useState } from 'react';
+import React, { FC, useMemo } from 'react';
 
 import classNames from 'classnames';
 import { t } from 'i18next';
@@ -8,23 +8,29 @@ import { Decimal } from '@sovryn/utils';
 
 import { AmountRenderer } from '../../../../../../2_molecules/AmountRenderer/AmountRenderer';
 import { BITCOIN, USD } from '../../../../../../../constants/currencies';
-import { useAccount } from '../../../../../../../hooks/useAccount';
 import { translations } from '../../../../../../../locales/i18n';
-import { getCurrencyPrecision } from '../../../ProtocolSection/ProtocolSection.utils';
-import { getConvertedValue } from '../../AssetSection.utils';
+import {
+  getCurrencyPrecision,
+  getConvertedValue,
+} from '../../ProtocolSection.utils';
 
-type TotalAssetSectionProps = {
+type ProtocolTotalSectionProps = {
   totalValue: Decimal;
+  selectedCurrency: string;
   btcPrice: string;
+  onCurrencyChange: (currency: string) => void;
+  className?: string;
+  title?: string;
 };
 
-export const TotalAssetSection: FC<TotalAssetSectionProps> = ({
+export const ProtocolTotalSection: FC<ProtocolTotalSectionProps> = ({
   totalValue,
+  selectedCurrency,
   btcPrice,
+  onCurrencyChange,
+  className,
+  title = t(translations.portfolioPage.protocolSection.totalValue),
 }) => {
-  const { account } = useAccount();
-  const [selectedCurrency, setSelectedCurrency] = useState(BITCOIN);
-
   const currencies = useMemo(() => [BITCOIN, USD], []);
 
   const renderCurrencyClassName = useMemo(
@@ -35,28 +41,17 @@ export const TotalAssetSection: FC<TotalAssetSectionProps> = ({
     [selectedCurrency],
   );
 
-  const convertedAmount = useMemo(
-    () =>
-      account
-        ? getConvertedValue(totalValue, selectedCurrency, btcPrice)
-        : Decimal.ZERO,
-    [btcPrice, selectedCurrency, totalValue, account],
-  );
-
   return (
-    <div className="flex flex-col gap-2 w-full md:max-w-[23.25rem] mb-2.5">
+    <div className={classNames(className, 'flex flex-col gap-3')}>
       <div className="flex justify-between items-center">
-        <Paragraph className="text-gray-30 font-medium">
-          {t(translations.portfolioPage.assetSection.totalAssets)}
-        </Paragraph>
-
+        <Paragraph className="text-gray-30 font-medium" children={title} />
         <div className="flex justify-end">
           {currencies.map(currency => (
             <Button
               key={currency}
               text={t(translations.common[currency.toLowerCase()])}
               className={renderCurrencyClassName(currency)}
-              onClick={() => setSelectedCurrency(currency)}
+              onClick={() => onCurrencyChange(currency)}
               style={ButtonStyle.secondary}
               dataAttribute={`portfolio-total-value-${currency.toLowerCase()}`}
             />
@@ -64,9 +59,9 @@ export const TotalAssetSection: FC<TotalAssetSectionProps> = ({
         </div>
       </div>
 
-      <div className="rounded md:bg-gray-80 text-gray-10 font-medium md:text-[2.25rem] md:p-3 truncate">
+      <div className="rounded md:bg-gray-80 text-gray-10 font-medium md:text-[2.25rem] md:p-3 mb-4 truncate">
         <AmountRenderer
-          value={convertedAmount}
+          value={getConvertedValue(totalValue, selectedCurrency, btcPrice)}
           suffix={selectedCurrency}
           precision={getCurrencyPrecision(selectedCurrency)}
           dataAttribute="portfolio-total-value"
