+++ apps/frontend/src/locales/en/translations.json
@@ -1445,9 +1445,7 @@
             "deposit": "Deposit",
             "openLoan": "Open loan",
             "stakeSOV": "Stake SOV",
-            "depositValue": "Deposit value"
-        },
-        "protocolTvlSection": {
+            "depositValue": "Deposit value",
             "totalValue": "Total value",
             "totalRewardsEarned": "Total rewards earned"
         },
